Rename legacy_global -> legacy
